refactor(engine): make output reply strategies a detail of the batch engine
