--- alchemist-engine/src/main/java/it/unibo/alchemist/core/model/OutputReplayStrategy.kt
@@ -1,39 +1,0 @@
-/*
- * Copyright (C) 2010-2023, Danilo Pianini and contributors
- * listed, for each module, in the respective subproject's build.gradle.kts file.
- *
- * This file is part of Alchemist, and is distributed under the terms of the
- * GNU General Public License, with a linking exception,
- * as described in the file LICENSE in the Alchemist distribution's top directory.
- */
-
-package it.unibo.alchemist.core.model
-
-/**
- * Engine output replay strategy modes.
- * @property code mode code
- */
-enum class OutputReplayStrategy(val code: String) {
-
-    /**
-     *  Launch simulation in aggregate output mode.
-     */
-    AGGREGATE("aggregate"),
-
-    /**
-     *  Launch simulation in replay output mode.
-     */
-    REPLAY("replay"), ;
-
-    companion object {
-        /**
-         * Parse string into OutputReplayStrategy.
-         */
-        fun parseCode(code: String): OutputReplayStrategy {
-            val match = OutputReplayStrategy.values().find { it.code == code }
-            return requireNotNull(match) {
-                "Unknown OutputReplayStrategy value $code, allowed: [${OutputReplayStrategy.values()}]"
-            }
-        }
-    }
-}
